feat: move near route component
